+++ tests/unit/framework/validators/ExistValidatorTest.php
@@ -36,18 +36,18 @@ class ExistValidatorTest extends DatabaseTestCase
 		}
 		// combine to save the time creating a new db-fixture set (likely ~5 sec)
 		try {
-			$val = new ExistValidator(['className' => ValidatorTestMainModel::className()]);
+			$val = new ExistValidator(['targetClass' => ValidatorTestMainModel::className()]);
 			$val->validate('ref');
 			$this->fail('Exception should have been thrown at this time');
 		} catch (Exception $e) {
 			$this->assertInstanceOf('yii\base\InvalidConfigException', $e);
-			$this->assertEquals('The "attributeName" property must be set.', $e->getMessage());
+			$this->assertEquals('The "attributeName" property must be configured as a string.', $e->getMessage());
 		}
 	}
 
 	public function testValidateValue()
 	{
-		$val = new ExistValidator(['className' => ValidatorTestRefModel::className(), 'attributeName' => 'id']);
+		$val = new ExistValidator(['targetClass' => ValidatorTestRefModel::className(), 'targetAttribute' => 'id']);
 		$this->assertTrue($val->validate(2));
 		$this->assertTrue($val->validate(5));
 		$this->assertFalse($val->validate(99));
@@ -57,22 +57,22 @@ class ExistValidatorTest extends DatabaseTestCase
 	public function testValidateAttribute()
 	{
 		// existing value on different table
-		$val = new ExistValidator(['className' => ValidatorTestMainModel::className(), 'attributeName' => 'id']);
+		$val = new ExistValidator(['targetClass' => ValidatorTestMainModel::className(), 'targetAttribute' => 'id']);
 		$m = ValidatorTestRefModel::find(['id' => 1]);
 		$val->validateAttribute($m, 'ref');
 		$this->assertFalse($m->hasErrors());
 		// non-existing value on different table
-		$val = new ExistValidator(['className' => ValidatorTestMainModel::className(), 'attributeName' => 'id']);
+		$val = new ExistValidator(['targetClass' => ValidatorTestMainModel::className(), 'targetAttribute' => 'id']);
 		$m = ValidatorTestRefModel::find(['id' => 6]);
 		$val->validateAttribute($m, 'ref');
 		$this->assertTrue($m->hasErrors('ref'));
 		// existing value on same table
-		$val = new ExistValidator(['attributeName' => 'ref']);
+		$val = new ExistValidator(['targetAttribute' => 'ref']);
 		$m = ValidatorTestRefModel::find(['id' => 2]);
 		$val->validateAttribute($m, 'test_val');
 		$this->assertFalse($m->hasErrors());
 		// non-existing value on same table
-		$val = new ExistValidator(['attributeName' => 'ref']);
+		$val = new ExistValidator(['targetAttribute' => 'ref']);
 		$m = ValidatorTestRefModel::find(['id' => 5]);
 		$val->validateAttribute($m, 'test_val_fail');
 		$this->assertTrue($m->hasErrors('test_val_fail'));
@@ -88,7 +88,7 @@ class ExistValidatorTest extends DatabaseTestCase
 		$val->validateAttribute($m, 'a_field');
 		$this->assertTrue($m->hasErrors('a_field'));
 		// check array
-		$val = new ExistValidator(['attributeName' => 'ref']);
+		$val = new ExistValidator(['targetAttribute' => 'ref']);
 		$m = ValidatorTestRefModel::find(['id' => 2]);
 		$m->test_val = [1,2,3];
 		$val->validateAttribute($m, 'test_val');
@@ -98,8 +98,8 @@ class ExistValidatorTest extends DatabaseTestCase
 	public function testValidateCompositeKeys()
 	{
 		$val = new ExistValidator([
-			'className' => OrderItem::className(),
-			'attributeName' => ['order_id', 'item_id'],
+			'targetClass' => OrderItem::className(),
+			'targetAttribute' => ['order_id', 'item_id'],
 		]);
 		// validate old record
 		$m = OrderItem::find(['order_id' => 1, 'item_id' => 2]);
@@ -115,8 +115,8 @@ class ExistValidatorTest extends DatabaseTestCase
 		$this->assertTrue($m->hasErrors('order_id'));
 
 		$val = new ExistValidator([
-			'className' => OrderItem::className(),
-			'attributeName' => ['order_id', 'item_id' => 2],
+			'targetClass' => OrderItem::className(),
+			'targetAttribute' => ['id' => 'order_id'],
 		]);
 		// validate old record
 		$m = Order::find(1);